+++ src/cli/dev.ts
@@ -1,0 +1,14 @@
+/**
+ * ARGS:
+ * port
+ *
+ */
+
+// import application
+// get list of commands
+// wipe commands on dev server
+// register commands on dev server
+// start application server in dev mode
+// watch for command changes and re-register them as needed
+// watch for changes to application itself and copy command list, restart the server, and compare new commands
+// restart notice: watch for changes to dotenv
